Fix issue#4, "flake8 not found" OS X #4
flake8 extension always complain about "flake8 not found", this is due to flake8 by default is not in the system path of OS X, so change this setting default to false will get rid of the pain of Mac OS X user. And pep8 extension set the pep8InSystemPath = false too.